Added parties & categories to FTT views
Redesigned display_categories helper

Since all the helpers are available in all the views and I've reused the
display_categories helper from EatDecisionsHelper in FttDecisionsHelper,
when the examples were run, the helper from EatDecisionsHelper were
being picked up first so the examples were failing.

That's why I created the generic_helper.rb and added the helper method
there, so that both models can use it.

The only drawback of the method is that it will fail for non CamelCase
models such as Decision, whereas EatDecision is fine.

subcategory is already defined